backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Hydra Orchestrator - Meta-agent system for coordinating Head agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.1",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
]
